zavrseno t4

+++ project/src/main/java/com/iktpreobuka/project/controllers/CategoryController.java
@@ -3,6 +3,7 @@ package com.iktpreobuka.project.controllers;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -13,7 +14,9 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.iktpreobuka.project.entities.CategoryEntity;
+import com.iktpreobuka.project.services.BillService;
 import com.iktpreobuka.project.services.CategoryService;
+import com.iktpreobuka.project.services.OfferService;
 
 
 
@@ -22,6 +25,12 @@ import com.iktpreobuka.project.services.CategoryService;
 public class CategoryController {
 	@Autowired
     private CategoryService categoryService;
+	
+	@Autowired
+	private OfferService offerService;
+	
+	@Autowired
+	private BillService billService;
 
     @GetMapping
     public List<CategoryEntity> getAllCategories() {
@@ -34,8 +43,16 @@ public class CategoryController {
     }
 
     @DeleteMapping("/{id}")
-    public void deleteCategory(@PathVariable Integer id) {
+    public ResponseEntity<?> deleteCategor(@PathVariable Integer id) {
+    	if(offerService.existsActiveOffersForCategory(id)) {
+            return ResponseEntity.badRequest().body("Cannot delete category with active offers.");
+        }
+        if(billService.existsActiveBillsForCategory(id)) {
+            return ResponseEntity.badRequest().body("Cannot delete category with active bills.");
+        }
+        // Proces brisanja kategorije ako nema aktivnih ponuda ili računa
         categoryService.deleteCategory(id);
+        return ResponseEntity.ok().build();
     }
 
     @GetMapping("/{id}")
